get cut3d tool to work again
